feat: applied main route
