+++ src/main.jsx
@@ -1,7 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
+import { BrowserRouter } from 'react-router-dom'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <App />
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
 )
