+++ src/routes/RoutesMain.jsx
@@ -1,0 +1,10 @@
+import { Route, Routes } from "react-router-dom"
+import { HomePage } from "../pages"
+
+export const RoutesMain = () => {
+    return (
+        <Routes>
+            <Route path="/"  element={<HomePage />} />
+        </Routes>
+    )
+}
